+++ flink-connector-oceanbase-directload/src/main/java/com/oceanbase/connector/flink/table/OceanBaseRowDataSerializationSchema.java
