+++ flink-connector-oceanbase-directload/src/test/resources/log4j2-test.properties
