+++ flink-connector-oceanbase-directload/src/test/resources/sql/mysql/products.sql
